UNI-77, UNI-105, UNI-104, UNI-99, UNI-74 institutions, account and player profile APIs, argon2

+++ internal/domain/institution.go
@@ -23,7 +23,7 @@ type InstitutionFilter struct {
 	City string
 }
 
-func NewInstitutionFilter(city *string, itype *int32) InstitutionFilter {
+func NewInstitutionFilter(city *string, itype *int8) InstitutionFilter {
 	f := InstitutionFilter{}
 
 	if itype != nil {
